test(runtime): separate eval scorecard fixtures
